+++ src/scripts/generate_theme.ts
@@ -4,9 +4,9 @@
 
 import fs from 'fs';
 import { buildColorSchemeMd } from '../theme_builder/color_scheme_md';
-import { DIST_DIR, DIST_MDSTYLE_CSS_CONTRIB_PATH, DIST_MDSTYLE_CSS_THEMED_PATH, DIST_MDSTYLE_DIR, DIST_THEMES_DIR } from '../common_defs';
-import { buildMdstyleCss } from '../theme_builder/mdstyle';
-import { buildThemeJson, createAllThemes } from '../theme_builder/theme';
+import { DIST_DIR, DIST_MDSTYLE_CSS_CONST_THEMED_PATH, DIST_MDSTYLE_CSS_CONTRIB_PATH, DIST_MDSTYLE_CSS_THEMED_PATH, DIST_MDSTYLE_DIR, DIST_THEMES_DIR } from '../common_defs';
+import { buildMdstyleCss, buildMdstyleCssSingle } from '../theme_builder/mdstyle';
+import { buildThemeJson, createAllConstThemes, createDarkTheme } from '../theme_builder/theme';
 import { defaultConfig } from '../config';
 
 
@@ -17,19 +17,26 @@ for (const dir of [DIST_DIR, DIST_THEMES_DIR, DIST_MDSTYLE_DIR]) {
 }
 
 const cfg = defaultConfig();
-const themes = createAllThemes(cfg);
-for (const theme of themes) {
-    const json = buildThemeJson(theme);
-    fs.writeFile(theme.themeDistPath, json, (err) => {
+const theme = createDarkTheme(cfg);
+const themes = createAllConstThemes();
+for (const t of [...themes, theme]) {
+    const json = buildThemeJson(t);
+    fs.writeFile(t.themeDistPath, json, (err) => {
         if (err) throw err;
     });
 
-    fs.writeFile('./ColorScheme_' + theme.fileName + '.md', buildColorSchemeMd(theme), (err) => {
+    fs.writeFile('./ColorScheme_' + t.fileName + '.md', buildColorSchemeMd(t), (err) => {
         if (err) throw err;
     });
 }
 
-const css = buildMdstyleCss(themes)
+const const_css = buildMdstyleCss(themes)
+fs.writeFile(DIST_MDSTYLE_CSS_CONST_THEMED_PATH, const_css, (err) => {
+    if (err) throw err;
+});
+
+const css = const_css + "\n" + buildMdstyleCssSingle(theme);
+
 fs.writeFile(DIST_MDSTYLE_CSS_THEMED_PATH, css, (err) => {
     if (err) throw err;
 });